refactor: move button component with topbar project
