+++ src/components/organisms/topbar/project/projectConfig.tsx
@@ -1,21 +1,17 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 
 import { useTranslation } from "react-i18next";
 import { redirect, useParams } from "react-router-dom";
 
 import { ProjectsService } from "@services";
-import { DrawerName } from "@src/enums/components";
 import { Project } from "@type/models";
 import { cn } from "@utilities";
 
 import { useFileOperations } from "@hooks";
-import { useDrawerStore, useProjectStore, useToastStore } from "@store";
+import { useProjectStore, useToastStore } from "@store";
 
-import { Button, ErrorMessage, IconSvg } from "@components/atoms";
+import { ErrorMessage } from "@components/atoms";
 import { ProjectTopbarButtons } from "@components/organisms/topbar/project";
-
-import { RunIcon } from "@assets/image";
-import { GearIcon } from "@assets/image/icons";
 
 export const ProjectConfigTopbar = () => {
 	const { t } = useTranslation(["projects", "errors", "buttons"]);
@@ -24,8 +20,6 @@ export const ProjectConfigTopbar = () => {
 	const [isNameValid, setIsNameValid] = useState<boolean>(true);
 	const [project, setProject] = useState<Project>();
 	const addToast = useToastStore((state) => state.addToast);
-	const { openDrawer } = useDrawerStore();
-
 	const inputClass = cn(
 		"min-w-3 rounded bg-transparent p-0 text-xl font-bold leading-6 leading-tight outline outline-0",
 		{
@@ -110,11 +104,6 @@ export const ProjectConfigTopbar = () => {
 		setIsNameValid(validateName(newName));
 	};
 
-	const openManualRunSettings = useCallback(() => {
-		openDrawer(DrawerName.projectManualRunSettings);
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, []);
-
 	return (
 		<div className="flex items-center justify-between gap-5 rounded-b-xl bg-gray-1250 py-2 pl-7 pr-3">
 			<div className="flex items-center gap-5">
@@ -138,28 +127,6 @@ export const ProjectConfigTopbar = () => {
 					</ErrorMessage>
 
 					<span className="text-sm font-semibold leading-tight">{project?.id}</span>
-				</div>
-
-				<div className="border-1 flex h-8 gap-2 rounded-3xl border border-gray-1000 p-1">
-					<Button
-						ariaLabel={t("topbar.buttons.ariaSettingsRun")}
-						className="h-full whitespace-nowrap p-1"
-						onClick={openManualRunSettings}
-						title={t("topbar.buttons.ariaSettingsRun")}
-						variant="filledGray"
-					>
-						<IconSvg className="fill-white" src={GearIcon} />
-					</Button>
-
-					<Button
-						ariaLabel={t("topbar.buttons.manual")}
-						className="h-full whitespace-nowrap px-3.5"
-						variant="filledGray"
-					>
-						<IconSvg src={RunIcon} />
-
-						{t("topbar.buttons.manual")}
-					</Button>
 				</div>
 			</div>
